+++ components/archive/api/src/main/java/com/flipperdevices/archive/model/CategoryType.kt
@@ -1,11 +1,13 @@
 package com.flipperdevices.archive.model
 
 import android.os.Parcelable
+import androidx.compose.runtime.Immutable
 import com.flipperdevices.bridge.dao.api.model.FlipperKeyType
 import kotlinx.parcelize.Parcelize
 import kotlinx.serialization.Serializable
 
 @Serializable
+@Immutable
 sealed class CategoryType : Parcelable {
     @Parcelize
     @Serializable