+++ components/archive/category/src/main/java/com/flipperdevices/archive/category/model/CategoryNavigationConfig.kt
@@ -1,0 +1,14 @@
+package com.flipperdevices.archive.category.model
+
+import com.flipperdevices.archive.model.CategoryType
+import com.flipperdevices.bridge.dao.api.model.FlipperKeyPath
+import kotlinx.serialization.Serializable
+
+@Serializable
+sealed class CategoryNavigationConfig {
+    @Serializable
+    data class Category(val categoryType: CategoryType) : CategoryNavigationConfig()
+
+    @Serializable
+    data class OpenKey(val keyPath: FlipperKeyPath) : CategoryNavigationConfig()
+}