+++ components/bridge/dao/api/src/main/java/com/flipperdevices/bridge/dao/api/model/FlipperKey.kt
@@ -10,6 +10,7 @@ import kotlinx.serialization.Serializable
  * The most complete description of the key
  */
 @Parcelize
+@Immutable
 data class FlipperKey(
     val mainFile: FlipperFile,
     val additionalFiles: List<FlipperFile> = emptyList(),