+++ components/infrared/impl/src/main/kotlin/com/flipperdevices/infrared/impl/api/InfraredViewDecomposeComponentImpl.kt
@@ -19,6 +19,7 @@ import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 
+@Suppress("LongParameterList")
 class InfraredViewDecomposeComponentImpl @AssistedInject constructor(
     @Assisted componentContext: ComponentContext,
     @Assisted private val keyPath: FlipperKeyPath,