+++ components/share/receive/src/main/java/com/flipperdevices/share/receive/api/ArchiveKeyDeeplinkHandlerImpl.kt
@@ -1,10 +1,6 @@
 package com.flipperdevices.share.receive.api
 
-import android.content.Intent
-import androidx.core.net.toUri
 import androidx.navigation.NavController
-import com.flipperdevices.archive.api.ArchiveFeatureEntry
-import com.flipperdevices.bottombar.api.BottomNavigationHandleDeeplink
 import com.flipperdevices.bridge.dao.api.model.FlipperFilePath
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.deeplink.api.DeepLinkHandler
@@ -13,7 +9,6 @@ import com.flipperdevices.deeplink.model.Deeplink
 import com.flipperdevices.deeplink.model.DeeplinkContent
 import com.flipperdevices.metric.api.MetricApi
 import com.flipperdevices.metric.api.events.SimpleEvent
-import com.flipperdevices.share.api.KeyReceiveFeatureEntry
 import com.squareup.anvil.annotations.ContributesBinding
 import com.squareup.anvil.annotations.ContributesMultibinding
 import javax.inject.Inject
@@ -21,10 +16,7 @@ import javax.inject.Inject
 @ContributesBinding(AppGraph::class)
 @ContributesMultibinding(AppGraph::class, DeepLinkHandler::class)
 class ArchiveKeyDeeplinkHandlerImpl @Inject constructor(
-    private val metricApi: MetricApi,
-    private val archiveFeatureEntry: ArchiveFeatureEntry,
-    private val keyReceiveFeatureEntry: KeyReceiveFeatureEntry,
-    private val bottomHandleDeeplink: BottomNavigationHandleDeeplink
+    private val metricApi: MetricApi
 ) : DeepLinkHandler {
     override fun isSupportLink(link: Deeplink): DispatcherPriority? {
         var content: DeeplinkContent? = null
@@ -51,17 +43,5 @@ class ArchiveKeyDeeplinkHandlerImpl @Inject constructor(
 
     override fun processLink(navController: NavController, link: Deeplink) {
         metricApi.reportSimpleEvent(SimpleEvent.OPEN_SAVE_KEY)
-
-        val archiveIntent = Intent().apply {
-            data = archiveFeatureEntry.getArchiveScreenByDeeplink().toUri()
-        }
-
-        val keyReceiveIntent = Intent().apply {
-            data = keyReceiveFeatureEntry.getKeyReceiveScreenDeeplinkUrl(link).toUri()
-            flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-        }
-
-        bottomHandleDeeplink.handleDeepLink(archiveIntent)
-        navController.handleDeepLink(keyReceiveIntent)
     }
 }