+++ components/widget/api/src/main/java/com/flipperdevices/widget/api/WidgetDecomposeComponent.kt
@@ -1,0 +1,13 @@
+package com.flipperdevices.widget.api
+
+import com.arkivanov.decompose.ComponentContext
+import com.flipperdevices.ui.decompose.DecomposeComponent
+
+interface WidgetDecomposeComponent : DecomposeComponent {
+    fun interface Factory {
+        operator fun invoke(
+            componentContext: ComponentContext,
+            widgetId: Int
+        ): WidgetDecomposeComponent
+    }
+}