+++ components/widget/screen/src/main/java/com/flipperdevices/widget/screen/model/WidgetNavigationConfig.kt
@@ -1,0 +1,13 @@
+package com.flipperdevices.widget.screen.model
+
+import com.flipperdevices.archive.api.SelectKeyPathListener
+import kotlinx.serialization.Serializable
+
+@Serializable
+sealed class WidgetNavigationConfig {
+    @Serializable
+    data class WidgetOptions(val widgetId: Int) : WidgetNavigationConfig()
+
+    @Serializable
+    data class SearchScreen(val listener: SelectKeyPathListener) : WidgetNavigationConfig()
+}