+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [Refactor] Migrate options screen to decompose
 - [Refactor] Migrate device screen to decompose
 - [Refactor] Migrate hub screen to decompose
+- [Refactor] Migrate archive screen to decompose
 - [Feature] New report bug
 
 # 1.6.7